requirements.txt
exifread
progressbar2
